// src/modules/auth/auth.interface.ts
export interface GoogleProfile {
  id?: string;
  displayName?: string;
  name?: {
    familyName?: string;
    givenName?: string;
  };
  emails?: {
    value?: string;
    verified?: boolean;
  }[];
  photos?: {
    value?: string;
  }[];
  provider: string;
}

export type JwtPayload = {
  sub: string;
  email: string;
};